refactor: separate logic hooks
